// server/types.ts
export interface ExchangeMarket {
  id: number
  symbol: string
  pair: string
  metadata: ExchangeMarketMetaData | null
  status: boolean
}

export interface ExchangeMarketMetaData {
  symbol: string
  base: string
  quote: string
  precision: Precision
  limits: Limits
  taker: number
  maker: number
}
// ...
type Precision = {
  amount: number
  price: number
}

type Limits = {
  amount: {
    min: number
    max: number
  }
  price: Record<string, any>
  cost: {
    min: number
    max: number
  }
  leverage: Record<string, any>
}